+++ aepp/fieldgroupmanager.py
@@ -58,6 +58,9 @@ class FieldGroupManager:
             full : OPTIONAL : Capability to force the full definition to be downloaded or not
             localFolder : OPTIONAL : If you want to use local storage to create all the connections between schema and field groups, classes and datatypes
             sandbox : OPTIONAL : If you use localFolder, you can specific the sandbox.
+        possible kwargs:
+            tenantId : OPTIONAL : If you want to specific the tenantId for the field group manager (if not provided, it will be retrieved from the schemaAPI or the local folder)
+            retry : int to set the number of retry in case of connection error for the schema module and the modules (default is the retry number set for the instance)
         """
         self.EDITABLE = False
         self.localfolder = None
@@ -66,11 +69,12 @@ class FieldGroupManager:
         self.dataTypes = {}
         self.dataTypeManagers = {} 
         self.requiredFields = set()
+        self.retry = kwargs.get("retry", aepp.config.config_object.get("retry",1))
         self.metaExtend = None
         if schemaAPI is not None and type(schemaAPI) == Schema:
             self.schemaAPI = schemaAPI
         elif config is not None and localFolder is None:
-            self.schemaAPI = Schema(config=config)
+            self.schemaAPI = Schema(config=config,retry=self.retry)
         elif localFolder is not None:
             if isinstance(localFolder, str):
                 self.localfolder = [Path(localFolder)]
@@ -241,7 +245,7 @@ class FieldGroupManager:
                 if self.schemaAPI is not None:
                     for dt in dataTypeManager_ids:
                         if dt not in self.dataTypes.keys():
-                            dt_manager = self.schemaAPI.DataTypeManager(dt)
+                            dt_manager = self.schemaAPI.DataTypeManager(dt,retry=self.retry)
                             self.dataTypes[dt_manager.id] = dt_manager.title
                             self.dataTypeManagers[dt_manager.title] = dt_manager
                 elif self.localfolder is not None: